Move DataRegion and DataSide under /data
