+++ devito/data/__init__.py
@@ -1,3 +1,4 @@
+from devito.data.meta import *  # noqa
 from devito.data.allocators import *  # noqa
 from devito.data.decomposition import *  # noqa
 from devito.data.data import *  # noqa
