+++ devito/dimension.py
@@ -5,10 +5,11 @@ from sympy.core.cache import cacheit
 import numpy as np
 from cached_property import cached_property
 
+from devito.data import LEFT, RIGHT
 from devito.exceptions import InvalidArgument
-from devito.types import LEFT, RIGHT, AbstractSymbol, Scalar
 from devito.logger import debug
 from devito.tools import ArgProvider, Pickable, dtype_to_cstr
+from devito.types import AbstractSymbol, Scalar
 
 __all__ = ['Dimension', 'SpaceDimension', 'TimeDimension', 'DefaultDimension',
            'SteppingDimension', 'SubDimension', 'ConditionalDimension', 'dimensions']
